move simulator components into their own folder
