+++ src/components/simulator/canvasCore.js
