+++ src/components/simulator/coreContainer.css
