+++ src/components/simulator/coreContainer.js
