+++ src/components/simulator/coreInput.css
