+++ src/components/simulator/coreInput.js
