+++ src/components/simulator/coreLocation.css
