+++ src/components/simulator/finishRoundControl.js
