+++ src/components/simulator/playPauseControl.js
