+++ src/components/simulator/resetControl.js
