+++ src/components/simulator/runProgressIndicator.js
