+++ src/components/simulator/simulatorControls.css
