+++ src/components/simulator/simulatorControls.js
