+++ src/components/simulator/speedControl.js
