+++ src/helpers/coreEnums.js
